requirements.txt
numpy==1.13.1
matplotlib==2.0.2
jupyter==1.0.0
Pillow==4.1.1
tqdm==4.19.5
